backend/requirements.txt
numpy>=1.24.0
opencv-python>=4.7.0
pillow>=9.5.0
flask>=2.3.0
flask-cors>=3.0.10
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0 